prior predictive sims for CWR model

+++ run-cwr-sir.R
@@ -6,7 +6,7 @@ library(posterior)
 FNAME = 'epi-timeseries-fed-agg-CA-partial'
 SAVE_CSV = TRUE
 PLOT_TS = TRUE
-SIMULATE = FALSE
+SIMULATE = TRUE
 
 dat_org <- read_csv('joined_datasets_california2/epi-timeseries-fed-agg-CA-partial.csv')
 
@@ -17,32 +17,66 @@ dat_org <- read_csv('joined_datasets_california2/epi-timeseries-fed-agg-CA-parti
 #  facet_wrap(~name, scales = 'free_y')
 
 
+if (!SIMULATE) {
+  stan_dat <- list(
+    max_t = nrow(dat_org),
+    ts = 1:nrow(dat_org),
+    yw = dat_org$Staff.Active,
+    yr = dat_org$Residents.Active,
+    yc = dat_org$Community.Active,
+    N = dat_org$Residents.Population,
+    worker_pop = dat_org$Staff.Population[1],
+    state_pop = 39000000,
+    alpha = 1/14
+  )
 
-stan_dat <- list(
-  max_t = nrow(dat_org),
-  ts = 1:nrow(dat_org),
-  yw = dat_org$Staff.Active,
-  yr = dat_org$Residents.Active,
-  yc = dat_org$Community.Active,
-  N = dat_org$Residents.Population,
-  worker_pop = dat_org$Staff.Population[1],
-  state_pop = 39000000,
-  alpha = 1/14
-)
+  print( 'got to here' )
+  print(stan_dat)
 
-print( 'got to here' )
-print(stan_dat)
+  exec <- cmdstan_model(
+    'stan-scripts/fit-sir-cwr.stan',
+    include_path = paste0(getwd(), '/stan-scripts')
+  )
 
-exec <- cmdstan_model(
-  'stan-scripts/fit-sir-cwr.stan',
-  include_path = paste0(getwd(), '/stan-scripts')
-)
+  fit <- exec$sample(data = stan_dat, adapt_delta = 0.8)
 
-fit <- exec$sample(data = stan_dat, adapt_delta = 0.8)
-post <- as_draws_df(fit$draws())
+  post <- as_draws_df(fit$draws())
 
-if (SAVE_CSV) {
-  write_csv(post, paste0('stan-fits/fit-sir-cwr-', FNAME, '-', Sys.Date(), '.csv'))
+  if (SAVE_CSV) {
+    write_csv(post, paste0('stan-fits/fit-sir-cwr-', FNAME, '-', Sys.Date(), '.csv'))
+  }
+} else {
+  stan_dat <- list(
+    max_t = nrow(dat_org),
+    ts = 1:nrow(dat_org),
+    inf_init_res = dat_org$Residents.Active[1],
+    pop_init = dat_org$Residents.Population[1],
+    inf_init_worker = dat_org$Staff.Active[1],
+    inf_init_state = dat_org$Community.Active[1],
+    worker_pop = dat_org$Staff.Population[1],
+    state_pop = 39000000,
+    alpha = 1/14
+  )
+
+  # compile and fit the model
+  exec <- cmdstan_model(
+    'stan-scripts/simulate-sir-cwr.stan',
+    include_path = paste0(getwd(), '/stan-scripts')
+  )
+
+  fit <- exec$sample(
+    data = stan_dat,
+    chains = 1,
+    iter_sampling = 1000,
+    fixed_param = TRUE
+  )
+  print('finished fit')
+    post <- as_draws_df(fit$draws())
+
+  if (SAVE_CSV) {
+    date <- str_replace(Sys.time(), ':', '-')
+    write_csv(post, paste0('stan-fits/simulate-sir-cwr-', date, '.csv'))
+  }
 }
 
 if (PLOT_TS) {
@@ -61,7 +95,7 @@ if (PLOT_TS) {
     gg + geom_point(data = tibble(time = stan_dat$ts, inf = stan_dat$yr))
   else
     gg
-  ggsave("sir-cwr-ca-fed-agg-res.png")
+  ggsave("prior-p-sir-cwr-ca-fed-agg-res.png")
 }
 
 if (PLOT_TS) {
@@ -80,7 +114,7 @@ if (PLOT_TS) {
     gg + geom_point(data = tibble(time = stan_dat$ts, inf = stan_dat$yw))
   else
     gg
-  ggsave("sir-cwr-ca-fed-agg-worker.png")
+  ggsave("prior-p-sir-cwr-ca-fed-agg-worker.png")
 }
 
 if (PLOT_TS) {
@@ -99,7 +133,7 @@ if (PLOT_TS) {
     gg + geom_point(data = tibble(time = stan_dat$ts, inf = stan_dat$yc))
   else
     gg
-  ggsave("sir-cwr-ca-fed-agg-state.png")
+  ggsave("prior-p-sir-cwr-ca-fed-agg-state.png")
 }
